properly clean up removed widgets from the database

+++ app/src/main/kotlin/com/simplemobiletools/notes/pro/models/Widget.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "widgets", indices = [(Index(value = ["widget_id"], unique = true))])
 data class Widget(
-        @PrimaryKey(autoGenerate = true) var id: Int?,
+        @PrimaryKey(autoGenerate = true) var id: Long?,
         @ColumnInfo(name = "widget_id") var widgetId: Int,
         @ColumnInfo(name = "note_id") var noteId: Long)
